Add exponential filter matrix
This matrix is intended for filter stabilization.
This adds version for precomputed_spectral_quantity taking an instantiated
generator, so that filter options can be propagated down.

There are functions that generate a cache for the filter matrix
and functions that retrieve these cached functions.
The precomputed_spectral_quantity has a new overload to
make it possible to propagate down additional parameters, like
here for the parameters of the spectral filter.